Begin working on site minimizer
I want to build a tool that I can run as a Git post-receive hook which
minimizes all html files in a given tree, compresses them with Brotli,
and puts that somewhere, possibly checking it out.

This way I can push to a gh-pages branch like I already do, and have
this tool minimize everything, and then put it in the webroot so I can
deploy it from there.

At first I thought I'd just copy my own minimizer from my blog, but then
I started looking at Git interaction in Haskell, it appears pretty
limited, then I looked into FFI'ing libgit2, but that is going to be a
pretty big endeavour by itself. And on top of that, Haskell is difficult
to compile for 32-bit ARM, should I want to run this on an older
Raspberry Pi. So for now, let's see if we can take an off-the-shelf
minimizer in Rust, and patch it together with some libgit2.