Add a cache for storing compressed objects
Or well, the objects are stored in the Git database, but we store the
mapping of unminified blob to the various minified and compressed blobs
in a tsv file, so we can locate those blobs later.

Currently the blobs could be garbage-collected, but the plan is to build
a Git tree, and then a commit, and then a ref, which would ensure that
the blobs are not dangling.