Build a static binary with Nix
After fiddling with this for SO LONG, FINALLY I got it to work! Hooray!

The trick was to use pkgsStatic everywhere, but that was not enough, we
needed to pass a few -l flags to rustc to make it link the right
libraries. Oh and we need to override libgit2 to not depend on
http-parser.